Bug: Swerve gets stuck
When a vehicle is facing east (0 degrees) and it swerves left twice this
ran into a problem with angular degree comparison. One number I'm
comparing and up valued at 330 instead of -30, so both the distance from
zero and the sign end up different than I'd expect.

For unknown reasons, the same thing did not happen with bend.

Dumb, easy fix for now. Plus comments so that I remember it's weird.

In the future, I should have a general way to do angular comarisons that
don't run into this problem.