Bug: Swerve gets stuck
When a vehicle is facing east (0 degrees) and it swerves left twice this
ran into a problem with angular degree comparison. One number I'm
comparing and up valued at 330 instead of -30, so both the distance from
zero and the sign end up different than I'd expect.

For unknown reasons, the same thing did not happen with bend.

Added a degree diffing function that does the right thing so that I
won't have to think about it next time.