Subphases!
Broke phases into subphases, composed either of:
  - serial actions or
  - simultaneous reveal

They have modals for players to interact with. Both with mouse and
hotkeys.

To be able to use this, there is now godmode (?godmode=true) to switch
to another player's session, to see what they see and/or move them.

Lots of cleanup and code juggling. And horrible kludges.

This is ugly but seems to to work.

Next, make it easy for anyone to look at (other) cars.
Then add tests.

Then there are a slew of missing features or cleanup. Pick something
big-ish. Probably something in this order:
  1. run hosted (Heroku?)
  2. actual logins/accounts (Auth0?)
  3. collision resolution
  4. simple AI