Out of control - check it in and try smaller bites
This started with cleaning up and centralizing CSS.

Then I realized that that meant CSS for modals.

That led me to realize I was missing a state. And so I implemented speed
changes leading to possible handling problems and tire damage. One
piece of this is yet another modal! I need to rethink that.

And that led me to extending the color notification system for choosing
moves to also show up when changing speed. Hard braking has difficulty
and could lead to control rolls and crash table 1. And also to damage.
So I now try to warn with yellow/orange/red.

And that led me to fixing a few bugs.

And then I realized I should really add some tests. I was waiting to
determine the "shape" of this code before doing that but now it's time.

So I started adding those. And realized I'm way off track for the
initial goal of cleaning up CSS.

Thank goodness I'm still alone here or I'd feel obliged to cherry-pick
meaningful pieces of this apart into separate commits!

Next task: add server tests toward eventually adding tests for why we
now "stall" on skids, not allowing movement but prompting for it.