Refactor server code away from Apollo bindings
And fix some bugs.
And convert to TS.
And generally refactor.

There is still more to do. This is almost to the point where I can
convert server code to object-oriented so that it's easier to reason
about.

And to add some Mongo instead of that one big in-memory object.